refactor(codegen): slightly improve code quality
