+++ codegen/src/utils.rs
@@ -1,0 +1,10 @@
+#![macro_use]
+
+macro_rules! unwrap {
+    ($err:expr) => {
+        match $err {
+            Ok(res) => res,
+            Err(err) => return err.to_compile_error().into(),
+        }
+    };
+}
